+++ apps/console/env.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="next" />
-/// <reference types="next/image-types/global" />
-
-// SPDX-FileCopyrightText: 2026 OpenDX CompanyOS contributors
-// SPDX-License-Identifier: Apache-2.0